refactor: make DataAddressValidator extend Validator
