+++ core/control-plane/control-plane-aggregate-services/src/main/java/org/eclipse/edc/connector/service/dataaddress/DataAddressValidatorImpl.java
@@ -14,40 +14,41 @@
 
 package org.eclipse.edc.connector.service.dataaddress;
 
-import org.eclipse.edc.spi.dataaddress.DataAddressValidator;
-import org.eclipse.edc.spi.result.Result;
 import org.eclipse.edc.spi.types.domain.DataAddress;
 import org.eclipse.edc.spi.types.domain.HttpDataAddress;
+import org.eclipse.edc.validator.spi.DataAddressValidator;
+import org.eclipse.edc.validator.spi.ValidationResult;
 import org.jetbrains.annotations.NotNull;
 
 import java.net.MalformedURLException;
 import java.net.URL;
 
-import static java.lang.String.format;
 import static org.eclipse.edc.spi.types.domain.HttpDataAddress.HTTP_DATA;
+import static org.eclipse.edc.validator.spi.Violation.violation;
 
 public class DataAddressValidatorImpl implements DataAddressValidator {
 
     @Override
-    public Result<Void> validate(DataAddress dataAddress) {
+    public ValidationResult validate(DataAddress dataAddress) {
         if (HTTP_DATA.equals(dataAddress.getType())) {
             return validateHttpDataAddress(dataAddress);
         } else {
-            return Result.success();
+            return ValidationResult.success();
         }
     }
 
     @NotNull
-    private Result<Void> validateHttpDataAddress(DataAddress dataAddress) {
+    private ValidationResult validateHttpDataAddress(DataAddress dataAddress) {
         var httpDataAddress = HttpDataAddress.Builder.newInstance()
                 .properties(dataAddress.getProperties())
                 .build();
         var baseUrl = httpDataAddress.getBaseUrl();
         try {
             new URL(baseUrl);
-            return Result.success();
+            return ValidationResult.success();
         } catch (MalformedURLException e) {
-            return Result.failure(format("DataAddress of type %s must contain a valid baseUrl: %s", HTTP_DATA, baseUrl));
+            var violation = violation("DataAddress of type %s must contain a valid baseUrl: %s".formatted(HTTP_DATA, baseUrl), "baseUrl", baseUrl);
+            return ValidationResult.failure(violation);
         }
     }
 }
