+++ spi/common/validator-spi/src/main/java/org/eclipse/edc/validator/spi/DataAddressValidator.java
@@ -1,0 +1,24 @@
+/*
+ *  Copyright (c) 2022 Bayerische Motoren Werke Aktiengesellschaft (BMW AG)
+ *
+ *  This program and the accompanying materials are made available under the
+ *  terms of the Apache License, Version 2.0 which is available at
+ *  https://www.apache.org/licenses/LICENSE-2.0
+ *
+ *  SPDX-License-Identifier: Apache-2.0
+ *
+ *  Contributors:
+ *       Bayerische Motoren Werke Aktiengesellschaft (BMW AG) - initial API and implementation
+ *
+ */
+
+package org.eclipse.edc.validator.spi;
+
+import org.eclipse.edc.spi.types.domain.DataAddress;
+
+/**
+ * Validate a DataAddress
+ */
+public interface DataAddressValidator extends Validator<DataAddress> {
+
+}
